requirements.txt
Flask
openai-whisper
gunicorn
